fix(ROB-845): preserve MCP broker import isolation
